refactor(web): move reusable components to lib folder
